parameter-framework: Enabling remote control from Windows
BZ: 8740

Used cygwin environment to build remote-process exe and remote-processor library
Created also an installer

Reviewed-on: http://android.intel.com:8080/22313
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/26776

+++ remote-processor/Message.cpp
@@ -39,7 +39,7 @@ CMessage::CMessage(uint8_t ucMsgId) : _ucMsgId(ucMsgId), _pucData(NULL), _uiData
 {
 }
 
-CMessage::CMessage() : _ucMsgId(-1), _pucData(NULL), _uiDataSize(0), _uiIndex(0)
+CMessage::CMessage() : _ucMsgId((uint8_t)-1), _pucData(NULL), _uiDataSize(0), _uiIndex(0)
 {
 }
